pyproject.toml
[build-system]
requires = ["setuptools"]

[tool.setuptools]
packages = ["core"]

[project]
name = "core_utils"
authors = [
    {name = "Nathan Maxwell"}
]
description = "A library of various helpers that are inevitably rewritten in every project."
readme = "README.md"
requires-python = ">=3.10"
keywords = ["python", "core", "util", "utils", "helpers"]
classifiers = [
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    'importlib-metadata; python_version<"3.10"',
    "pytest",
    "pyyaml"
]
version = "0.5.1"

[project.urls]
Homepage = "https://github.com/nate-maxwell/core_utils"
Source = "https://github.com/nate-maxwell/core_utils"
